Add Point Scraper File and Move Point Code from Match File
